requirements.txt
numpy==1.12.1
pylint==1.7.1
pillow==4.0.0
logutils>=0.3
